File Movements
Still a lot to fix, like change all image names to actual freaking names that describe exactly what they are, reset a few pieces of the code so that its not trash, and re-add the main USSE.cs back in the game, but fixed. Requires referencing new libraries of course.